+++ CHANGELOG.md
@@ -21,6 +21,7 @@ https://github.com/nwnxee/unified/compare/build8193.21...HEAD
 ##### New NWScript Functions
 - Creature: Get{Armor|Shield}CheckPenalty()
 - Creature: {Get/Set}BypassEffectImmunity()
+- Creature: {Get/Set}LastKiller()
 - Effect: GetTrueEffectCount()
 - Effect: GetTrueEffect()
 - Effect: RemoveEffectById()